api/requirements.txt
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2
httptools==0.6.4
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# AI/LLM libraries
anthropic==0.64.0
langchain-anthropic==0.3.18
langchain-core==0.3.74
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14

# Database
SQLAlchemy==2.0.43
alembic==1.16.4
psycopg2-binary==2.9.10

# PDF Processing (keeping this as essential)
PyMuPDF==1.26.3
numpy==2.2.6

# Core utilities
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
requests-toolbelt==1.0.0
pillow==11.3.0

# HTTP and networking
httpcore==1.0.9
httpx==0.28.1
h11==0.16.0
anyio==4.10.0
sniffio==1.3.1
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0

# JSON and data handling
orjson==3.11.2
PyYAML==6.0.2
jsonpatch==1.33
jsonpointer==3.0.0
jiter==0.10.0
ormsgpack==1.10.0

# Template and security
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0

# Utilities
click==8.2.1
packaging==25.0
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.14.1
tenacity==9.1.2
distro==1.9.0
